cleanup structure and complete docs
