+++ publisher/src/cartography/cartographer.rs
@@ -1,83 +1,13 @@
 use async_trait::async_trait;
 use std::{ error::Error, sync::Arc, pin::Pin };
-use futures::stream::{ Stream, StreamExt };
+use futures::stream::Stream;
 use image::Pixel;
 
 use super::{ Storage, Coords, Point };
-use super::target::MapTarget;
-use super::schedule::{ Task, Scheduler };
-
-use log::info;
+use super::bitmask::{ target::BitmaskTarget, generate::{ GenerateAndStoreBitmaskTask, get_chunk_coords } };
+use super::schedule::Scheduler;
 
 
-const CHUNK_SIZE: u32 = 256;
-const BITMAP_SIZE: usize = (CHUNK_SIZE * CHUNK_SIZE) as usize / 8;
-
-///
-/// A task that generates a bitmask for a chunk of the map and stores it
-/// in the bitmask target.
-///
-#[derive(Clone)]
-pub struct GenerateAndStoreBitmaskTask<P: Pixel + Send + Sync + 'static> {
-  storage: Arc<dyn Storage<P>>,
-  target: Arc<dyn MapTarget>,
-}
-
-impl <P: Pixel + Send + Sync + 'static> GenerateAndStoreBitmaskTask<P> {
-  pub fn new(storage: Arc<dyn Storage<P>>, target: Arc<dyn MapTarget>) -> Self {
-    Self {
-      storage,
-      target,
-    }
-  }
-}
-
-#[async_trait]
-impl <P: Pixel + Send + Sync> Task for GenerateAndStoreBitmaskTask<P>
-where <P as Pixel>::Subpixel: Send + Sync {
-  async fn run(&self, start: &Coords) {
-    info!("Generating bitmask for coords {:?}", start);
-
-    //
-    // create a bitmask for the chunk
-    //
-    let mut bitmask = vec![0u8; BITMAP_SIZE];
-
-    //
-    // get the stored coordinates within the chunk
-    //
-    let end = (start.0 + CHUNK_SIZE as i32, start.1 + CHUNK_SIZE as i32);
-    let mut points = self.storage.get(&start, &end).await;
-
-    //
-    // set the corresponding bit of each point
-    //
-    while let Some(point) = points.next().await {
-
-      // the position of the point within the chunk
-      let pos = (point.coords.0 - start.0, point.coords.1 - start.1);
-
-      // the index of the bit in the bitmask
-      let bit_index = (pos.0 + pos.1 * CHUNK_SIZE as i32) as usize;
-
-      // the index of the byte containing target bit
-      let byte_index = bit_index / 8;
-
-      // the offset of the target bit within the byte
-      let bit_offset = bit_index % 8;
-
-      if byte_index < bitmask.len() {
-        // flip the bit to 1
-        bitmask[byte_index] |= 1 << bit_offset;
-      }
-    }
-
-    match self.target.save_bitmask(&start, &bitmask).await {
-      Ok(()) => info!("Bitmask for coords {:?} saved", start),
-      Err(err) => info!("Error saving bitmask for coords {:?}: {}", start, err),
-    }
-  }
-}
 
 ///
 /// A cartographer that schedules a bitmask generation task
@@ -101,7 +31,7 @@ where <P as Pixel>::Subpixel: Send + Sync {
   ///
   pub fn new(
     storage: Arc<dyn Storage<P>>,
-    target: Arc<dyn MapTarget>,
+    target: Arc<dyn BitmaskTarget>,
     mut scheduler: impl Scheduler<GenerateAndStoreBitmaskTask<P>> + 'static,
   ) -> Self {
     scheduler.assign(GenerateAndStoreBitmaskTask::new(Arc::clone(&storage), target));
@@ -123,7 +53,7 @@ where <P as Pixel>::Subpixel: Send + Sync {
   ) -> Result<(), Box<dyn Error + Send + Sync>> {
     match self.storage.put(coords, color).await {
       Ok(()) => {
-        let anchor = (coords.0 / (CHUNK_SIZE as i32), coords.1 / (CHUNK_SIZE as i32));
+        let anchor = get_chunk_coords(coords);
         Ok(self.scheduler.schedule(&anchor).await)
       },
       err => err
@@ -134,7 +64,7 @@ where <P as Pixel>::Subpixel: Send + Sync {
     -> Result<(), Box<dyn Error + Send + Sync>> {
       match self.storage.delete(coords).await {
         Ok(()) => {
-          let anchor = (coords.0 / (CHUNK_SIZE as i32), coords.1 / (CHUNK_SIZE as i32));
+          let anchor = get_chunk_coords(coords);
           Ok(self.scheduler.schedule(&anchor).await)
         },
         err => err
