+++ publisher/src/cartography/mod.rs
@@ -5,15 +5,14 @@ use sqlx::{ Pool, Sqlite };
 use image::Rgb;
 
 use storage::sqlite::SqliteMapStorage;
-use s3_target::S3MapTarget;
+use bitmask::s3_target::S3BitmaskTarget;
 use schedule::Throttler;
 use cartographer::Cartographer;
 
 pub mod storage;
 pub mod schedule;
 pub mod cartographer;
-pub mod target;
-pub mod s3_target;
+pub mod bitmask;
 pub use storage::{ Storage, Coords, Point };
 
 
@@ -21,7 +20,7 @@ pub type DefaultMapStorage = Cartographer<Rgb<u8>>;
 
 pub fn init(db: Pool<Sqlite>, s3: S3Client) -> impl Storage<Rgb<u8>> {
   let sqlite = SqliteMapStorage::new(db);
-  let target = S3MapTarget::new(s3, None);
+  let target = S3BitmaskTarget::new(s3, None);
 
   cartographer::Cartographer::new(
     Arc::new(sqlite),
